feat(gui): per-repo override for never_approve_with_issues
Lift _computeRepoDiff/_listsEqual out of RepoDetailScreen into a
top-level computeRepoDiff (repo_diff.dart) so the diff logic is
unit-testable, then add the never_approve_with_issues diff entry and
its OverrideDropdown, mirroring the existing pr_draft pattern.
